+++ lib/src/constants/enums.dart
@@ -1,0 +1,13 @@
+///
+/// [Author] Alex (https://github.com/Alex525)
+/// [Date] 2020/7/8 12:32
+///
+
+/// Provide some special picker types to integrate
+/// un-common pick pattern.
+/// 提供一些特殊的选择器类型以整合非常规的选择行为。
+enum SpecialPickerType { wechatMoment }
+
+/// Provide an item slot for custom widget insertion.
+/// 提供一个自定义位置供自定义item放入资源列表中。
+enum CustomItemPosition { none, prepend, append }