+++ lib/src/provider/asset_picker_viewer_provider.dart
@@ -1,5 +1,5 @@
 ///
-/// [Author] Alex (https://github.com/AlexVincent525)
+/// [Author] Alex (https://github.com/Alex525)
 /// [Date] 2020/3/31 15:37
 ///
 import 'package:flutter/material.dart';
@@ -14,7 +14,7 @@ class AssetPickerViewerProvider extends ChangeNotifier {
     _currentlySelectedAssets = List<AssetEntity>.from(assets);
   }
 
-  /// Selected assets in viewer.
+  /// Selected assets in the viewer.
   /// 查看器中已选择的资源
   List<AssetEntity> _currentlySelectedAssets;