+++ lib/wechat_assets_picker.dart
@@ -1,10 +1,10 @@
 library wechat_assets_picker;
 
-export 'package:photo_manager/photo_manager.dart'
-    show AssetEntity, RequestType, AssetType;
+export 'package:photo_manager/photo_manager.dart';
 
+export 'src/constants/enums.dart';
+export 'src/delegates/assets_picker_text_delegate.dart';
 export 'src/delegates/sort_path_delegate.dart';
-export 'src/delegates/text_delegate.dart';
 export 'src/provider/asset_entity_image_provider.dart';
 export 'src/provider/asset_picker_provider.dart';
 export 'src/provider/asset_picker_viewer_provider.dart';